requirements.txt
fastapi>=0.111.0
slowapi>=0.1.9
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9
anthropic>=0.34.0
python-dotenv>=1.0.0
pymupdf==1.24.5
pymupdf4llm==0.0.17
sentence-transformers==3.0.1
chromadb==0.5.3
numpy>=1.26.0
torch>=2.3.0
tqdm==4.66.4
tiktoken==0.7.0
